Add support for IPFS links to the parser.
I wonder if it should just handle anything that looks like a URL instead of
hard-coding a short list of protocols.  Incidentally, the link I was trying
to add to my personal wiki was ipfs://bafybeic5awfcndlo7uxaigkjhkop3fle3enegswsofo66hw3n5243enggi/ ,
which is a mirror of https://github.kdex.de/earthbound-battle-backgrounds/ ,
which is pretty cool.